requirements.txt
click
numpy
Pillow
torch==1.2.0
scipy
torchvision==0.4.0
horovod==0.16.4
